Test again

+++ Python/main.py
@@ -19,7 +19,7 @@ print('Reading serial data...')
 
 data = serial.read(BYTES_TO_READ)
 d_type = np.dtype(np.uint16).newbyteorder('<')
-np_data = np.frombuffer(data, dtype=d_type)
+np_data = np.frombuffer(data, dtype=d_type).astype(np.int16)
 
 np_data = np_data - 2048
